Merged PR 37917: Use Euclidean (again) not Haversine
Annotate queryset with distance from given coordinates using Euclidean formula.
Uses simplified distance calculation that treats Earth as flat plane.

More accurate calculation is the Haversine formula, but this is more complex
and requires more computational power.
With an average of ~11 meters difference between the two calculations,
within Amsterdam and a maximum of ~46 meters difference,
this is good enough for our use case.

The calculation gives a distance in kilometers.

Related work items: #132153