Merged PR 39954: Fix sending logs to App Insights via OpenTelemetry
It seems like the instrumention with Django was working, since we received the requests in Application Insights. But there were no logs visible.

This is a refactor of how logging is setup up. It doesn't have to be setup per service anymore, but is done through the "core app".

Related work items: #137475