Merged PR 40009: Implement cut off moment
It comes down to the new super().is_valid() function for tokens:

This function checks if token is passed its globally defined cut off datetime.
This datetime represents the moment when new "stadspas" data made available by the organisation.
This moment happens every year, so the format is "month-day hour:minute" or "%m-%d %H:%M"

If token is created after cut off datetime, it's valid.
If token is created before or on cut off datetime, and the current time is after cut off datetime,
the token is invalid and will be deleted.

Related work items: #137345