Merged PR 38623: Setup notification config logic
We're not yet actually checking if a device has a notification type enabled/disabled for push, since the frontend does not have this feature yet. This logic is present, but commented out for now.

Includes the config for the debug-toolbar, which makes it easy to check which SQL queries are being performed.

Also included in the core service is the DeviceIdMixin, only used by notifcation service for now.