Merged PR 43110: Unit test coverage check + unit test fixes
Affected services
- construction_work
- image

Definition of Done
- [ ] Swagger UI tested & up-to-date (`make dev`)
- [x] Dependencies updated (`make requirements`)
- [ ] Unittest coverage >90% (`make coverage`)
- [ ] API endpoints backwards compatible (`make openapi-diff`)
- [ ] Infrastructure repo updated (aapp_azure_infra)
- [ ] Loadtests for relevant endpoints (aapp_testing_loadtests)

After deployment
- [ ] Manual check on "Ontwikkel" en "Test" (e.g.: use feature via Swagger or "beheer" panel)
- [ ] Sanity check on development app

The main feature here is the addition of a unit test coverage check. You can run `make coverage` to check all services, or specify a service by prepending it with `SERVICE_NAME=xxx`. In the `pyproject.toml` the minimal coverage is set to 90%. As of now, only the Contact service does not match this limit, at 87%. Therefor it is not yet included in the `make test` run.

Additionally I noticed that some unit tests were creating image files locally. Turns out that the `DEFAULT_FILE_STORAGE` setting was deprecated since Django 4.2.

I also removed the settings and requirements for `black`, `isort` and `autoflake`, since this is now all covered by `ruff`.