Merged PR 43219: Session delete endpoint, xxx_date_time uniformity, empty list no 404
Affected services
- bridge (parking)

Definition of Done
- [x] Swagger UI tested & up-to-date (`make dev`)
- [ ] Dependencies updated (`make requirements`)
- [x] Unittest coverage >90% (`make coverage`)
- [ ] API endpoints backwards compatible (`make openapi-diff`)
- [ ] Infrastructure config updated (aapp_azure_infra)
- [ ] Loadtests for relevant endpoints (aapp_testing_loadtests)

After deployment
- [ ] Manual check on "Ontwikkel" en "Test" (e.g.: use feature via Swagger or admin panel)
- [ ] Sanity check on test app

Many changes to the API.

There is a new endpoint to delete a session with a HTTP DELETE call. This simplifies the action of deleting a session. Since the Egis logic for deleting a session is a PATCH with as end_date_time the current time. Now there is no need for the app to implement this.

It does break backwards compatibility for the license plate delete view, since from now on, a DELETE endpoint does not accept a body anymore, only query params.

The session receipt call now accepts `xxx_date_time` instead of `xxx_date` keys.

The session list response is now updated. Status values will be presented in the same way as the filter. The `xxx_date` fields are now presented as `xxx_date_time` for uniformity,

When a session list call has no results, a empty list in paginated format is returned, instead of a 404.

The `payment_zone_id` field is added to the session list result.