Merged PR 43419: Accept timestamps without capital T in string representation
Affected services
- bridge (parking)

Definition of Done:
- [X] API is backwards compatible (`make openapi-diff`)
- [x] Swagger UI up-to-date & tested (`make dev`)
- [ ] Dependencies updated (`make requirements`)
- [x] Unittest coverage >90% (`make coverage`)
- [ ] Infrastructure config updated (aapp_azure_infra)
- [ ] Loadtests for relevant endpoints (aapp_testing_loadtests)

If manually deployed:
- [ ] Manual check on dev (o) & test (t), e.g. use feature via Swagger or admin panel
- [ ] Sanity check on test app

Removes a string replace call. This also replaced a missing "+" char for the timezone offset. The replace was not needed to begin with since string with space between date and time is already a  valid ISO format.

It is important for the frontend to encode the query params, since a "+" char in a datetime is replaced with a space char, if string is not encoded.