Merged PR 43503: Return default result if selection key not found
Affected services
- bridge (parking)

Definition of Done:
- [ ] API is backwards compatible (`make openapi-diff`)
- [ ] Swagger UI up-to-date & tested (`make dev`)
- [ ] Dependencies updated (`make requirements`)
- [x] Unittest coverage >90% (`make coverage`)
- [ ] Infrastructure config updated (aapp_azure_infra)
- [ ] Loadtests for relevant endpoints (aapp_testing_loadtests)

If manually deployed:
- [ ] Manual check on dev (o) & test (t), e.g. use feature via Swagger or admin panel
- [ ] Sanity check on test app

This adds default behavior to a BaseSSPView that will return a dict or list (depending if response is many) value. This might either with default (empty) pagination or not. All of this is unit tested. Also other behavior of BaseSSPView is now properly tested and therefor covered.

Excluded unit tests from coverage check. This did lower the coverage of bridge from 94% to 91%.