Merged PR 43844: Increased test coverage Contact and Core
Affected services
- core
- contact

Definition of Done:
- [x] API is backwards compatible (`make openapi-diff`)
- [x] Swagger UI up-to-date & tested (`make dev`)
- [x] Dependencies updated (`make requirements`)
- [x] Unittest coverage >90% (`make coverage`)
- [ ] Infrastructure config updated (aapp_azure_infra)
- [ ] Loadtests for relevant endpoints (aapp_testing_loadtests)

If manually deployed:
- [ ] Manual check on dev (o) & test (t), e.g. use feature via Swagger or admin panel
- [ ] Sanity check on test app

The main purpose of this PR is to increase the test coverage. The lowest test coverage overall is now 93%. This is done by testing a bunch of core logic, also contact was missing some testing, especially the waiting times view, it was not tested at all, and had some sketchy code.

One odd thing is that there was a migration to be ran for the contact service. It seems like the previous manually created migration covers the needed changes in the database, but this was not recognized by Django. It has to be confirmed is this is going to work properly.

Some changes were made in the Makefile. When running `make test` the coverage check is now used, for both testing and coverage. By default the core service is always tested and cover checked, since every service will call the core in a slightly different way.