Merged PR 44214: Upload and host Iprox images via image service
Affected services
- core
- construction_work
- image

Definition of Done:
- [x] API is backwards compatible (`make openapi-diff`)
- [x] Swagger UI up-to-date & tested (`make dev`)
- [x] Dependencies updated (`make requirements`)
- [x] Unittest coverage >90% (`make coverage`)
- [ ] Infrastructure config updated (aapp_azure_infra)
- [X] Loadtests for relevant endpoints (aapp_testing_loadtests)

If manually deployed:
- [ ] Manual check on dev (o) & test (t), e.g. use feature via Swagger or admin panel
- [x] Sanity check on test app

The `get_image` function in `load_data.py` is replaced with the `post_image` function, which has the same input and output, but uploads the biggest image from `sources` via out Image service. The Image service has a new endpoint, so only the image URL has to be passed.

Not related to the changes, but I noticed that the `ImageVariantSerializer` has functions there were unreachable. `image` is set as `SerializerMethodField` which is read-only. So therefor the `validate_image` function is unreachable. Conversion from heic to jpg is already managed via `core.utils.image_utils.scale_image` function. So I removed those functions.