Merged PR 44973: Contact serving Django admin /w Entra token auth
Affected services
- core
- contact
- construction_work

Definition of Done:
- [ ] API is backwards compatible (`make openapi-diff`)
- [ ] Swagger UI up-to-date & tested (`make dev`)
- [ ] Dependencies updated (`make requirements`)
- [x] Unittest coverage >90% (`make coverage`)
- [ ] Infrastructure config updated (aapp_azure_infra)
- [ ] Loadtests for relevant endpoints (aapp_testing_loadtests)

If manually deployed:
- [x] Manual check on dev (o) & test (t), e.g. use feature via Swagger or admin panel
- [ ] Sanity check on test app

This enables the Django admin panel for the Contact service. Mozilla/Amsterdam OIDC is not yet implemented here yet, but can be added later. The same logic for validating the Entra token in construction work is used, and therefor moved to Core. The opening times models still need to be refactored.