Merged PR 45212: Opening hours model refactor (+ admin update)
Affected services
- contact
- core

Definition of Done:
- [x] API is backwards compatible (`make openapi-diff`)
- [x] Swagger UI up-to-date & tested (`make dev`)
- [x] Dependencies updated (`make requirements`)
- [x] Unittest coverage >90% (`make coverage`)
- [ ] Infrastructure config updated (aapp_azure_infra)
- [ ] Loadtests for relevant endpoints (aapp_testing_loadtests)

If manually deployed:
- [ ] Manual check on dev (o) & test (t), e.g. use feature via Swagger or admin panel
- [ ] Sanity check on test app

Starting of point is the refactor of the opening hour models. This results in many changes down the line:
- Admin is heavily updated, so opening hours can be easily managed by external users
- Migration to new models, where exception data is migrated from existing data (not reversible), and regular opening times is added via new `initregularhours` command (since these are almost static)
- Some minor changes to the views, serializers and tests

Access to admin panel is managed via Entra token. A cookie called `__Host-Access-Token` is expected in the browser which contains a token with the correct "my access group". This cookie is currently set by the "beheer" UI when navigating to "https://xxx.app.amsterdam.nl/beheer/contact".

There is a openapi-diff broken compatibilty alert on:
```
- GET    /contact/api/v1/city-offices
  Return Type:
    - Changed 200 OK
      Media types:
        - Changed application/json
          Schema: Broken compatibility
          Changed property type: result[n].visitingHours (string -> object)
```
This can be ignored, since I only added a spectacular decorator that renders correct OpenAPI specs for `visitingHours` instead of a string.