Merged PR 47704: Get scheduled not found as 204
Affected services
- core
- notification

Definition of Done:
- [ ] API is backwards compatible (`make openapi-diff`)
- [ ] Swagger UI up-to-date & tested (`make dev`)
- [ ] Dependencies updated (`make requirements`)
- [ ] Unittest coverage >90% (`make coverage`)
- [ ] Infrastructure config updated (aapp_azure_infra)
- [ ] Loadtests for relevant endpoints (aapp_testing_loadtests)

If manually deployed:
- [ ] Manual check on dev (o) & test (t), e.g. use feature via Swagger or admin panel
- [ ] Sanity check on test app

In the core `ScheduledNotificationService` we already accepting that on a `get()` call a 404 is acceptable, by returning `None`. But having a 404 in your happy flow is bad design, so this continue this logic only we return a 204 (No Content). This makes sure that the reminder_scheduler for parking maintains its logic, which is hard to restucture, but doesn't trigger a 404.